mainboard: Add Clevo W650SZ
This port is based on the code generated with autoport with Haswell
support (30890).

It can boot Arch Linux from mSATA or USB with SeaBIOS payload.

The EHCI debug port is at the right hand side of the laptop, but the
debug message stops to display after mrc.bin initializes USB.

Change-Id: I84340af9490d1a48a9e7e1e98fefc646ced1d923
Signed-off-by: Iru Cai <mytbk920423@gmail.com>